+++ test/fixture/flow-default.input.js
@@ -1,9 +1,0 @@
-/**
- * Very Important Transform
- */
-function veryImportantTransform(
-  input: Array<string>,
-  options: Object = {}
-): string {
-  return "42";
-}