+++ test/fixture/jsx.input.js
@@ -1,7 +1,0 @@
-var util = require('util');
-
-/**
- * This function returns the number one.
- * @returns {Number} numberone
- */
-module.exports = (<p>hello</p>);