+++ test/fixture/lint/lint.input.js
@@ -1,5 +1,8 @@
 /**
  * @param {String} foo bar
+ * @param {Array<Number>} foo bar
+ * @param {Array<Number>} foo bar
+ * @param {Array|Number} foo bar
  * @memberof notfound
  */