+++ test/fixture/lint/lint.output.js
@@ -1,6 +1,10 @@
         1:1  warning  could not determine @name for hierarchy
+        1:1  warning  could not determine @name for hierarchy
         2:1  warning  type String found, string is standard
-        3:1  warning  @memberof reference to notfound not found
-        7:1  warning  type String found, string is standard
+        3:1  warning  type Number found, number is standard
+        4:1  warning  type Number found, number is standard
+        5:1  warning  type Number found, number is standard
+        6:1  warning  @memberof reference to notfound not found
+       10:1  warning  type String found, string is standard
 
-⚠ 5 warnings
+⚠ 8 warnings