+++ test/fixture/nest_params.input.js
@@ -5,3 +5,15 @@
  */
 function foo(employee) {
 }
+
+/**
+ * @name foo
+ * @param {Object} options some options
+ * @param {number} options.much how much
+ * @param {number} bar something else
+ * @property {Object} theTime the current time
+ * @property {number} theTime.hours
+ * @property {number} theTime.minutes
+ * @property {number} theTime.seconds
+ * @returns {Object} foo something else
+ */